First implementation starting from svelte-kit starter kit
