+++ .github/workflows/publish-gh-pages.yaml
@@ -7,7 +7,7 @@ on:
             - main
 
 jobs:
-    build-deploy:
+    build_site:
         runs-on: ubuntu-latest
         defaults:
             run:
@@ -21,36 +21,35 @@ jobs:
               uses: actions/setup-node@v4
               with:
                   node-version: 20
+                  cache: 'npm'
 
-            - name: Get yarn cache directory path
-              id: yarn-cache-dir
-              run: echo "::set-output name=dir::$(yarn cache dir)"
-
-            - name: Cache yarn dependencies
-              uses: actions/cache@v3
-              id: yarn-cache
-              with:
-                  path: ${{ steps.yarn-cache-dir.outputs.dir }}
-                  key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
-                  restore-keys: |
-                      ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
-                      ${{ runner.os }}-yarn-
-
-            - name: Install yarn dependencies
-              # as we allow any keys we have to run the installer anyway
-              # if: steps.yarn-cache.outputs.cache-hit != 'true'
-              run: cd ..; yarn install
-
-            - name: Install yarn dependencies
-              # as we allow any keys we have to run the installer anyway
-              # if: steps.yarn-cache.outputs.cache-hit != 'true'
-              run: yarn install
+            - name: Install dependencies
+              run: npm install
 
             - name: Run Build
-              run: yarn run build
+              env:
+                BASE_PATH: '/${{ github.event.repository.name }}'
+              run: npm run build
 
-            - name: Upload Build
-              uses: JamesIves/github-pages-deploy-action@v4.4.3
+            - name: Upload Artifacts
+              uses: actions/upload-pages-artifact@v2
               with:
-                  branch: gh-pages # The branch the action should deploy to.
-                  folder: ./example-app/public/ # The folder the action should deploy.
+                # this should match the `pages` option in your adapter-static options
+                path: 'build/'
+      
+    deploy:
+          needs: build_site
+          runs-on: ubuntu-latest
+      
+          permissions:
+            pages: write
+            id-token: write
+      
+          environment:
+            name: github-pages
+            url: ${{ steps.deployment.outputs.page_url }}
+      
+          steps:
+            - name: Deploy
+              id: deployment
+              uses: actions/deploy-pages@v2
