+++ .gitignore
@@ -1,9 +1,11 @@
 .DS_Store
 node_modules
-example-app/node_modules/
-example-app/public/build/
-example-app/.DS_Store
-test.html
-
-dist
-example-app/public/favicon.psd
+/build
+/dist
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
