--- example-app/.gitignore
@@ -1,4 +1,0 @@
-/node_modules/
-/public/build/
-
-.DS_Store
