--- example-app/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "svelte-virtual-table-example-app",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "rollup": "^4.1.5",
-    "rollup-plugin-svelte": "^7.1.6",
-    "rollup-plugin-css-only": "^4.3.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^4.2.2"
-  },
-  "dependencies": {
-    "sirv-cli": "^2.0.2",
-    "svelte-virtual-table": "^1.1.3"
-  }
-}
