--- rollup.config.js
@@ -1,28 +1,0 @@
-import commonjs from '@rollup/plugin-commonjs';
-import resolve from '@rollup/plugin-node-resolve';
-import svelte from 'rollup-plugin-svelte';
-import pkg from './package.json' assert { type: "json" };
-
-const name = pkg.name
-	.replace(/^(@\S+\/)?(svelte-)?(\S+)/, '$3')
-	.replace(/^\w/, m => m.toUpperCase())
-	.replace(/-\w/g, m => m[1].toUpperCase());
-
-export default {
-	input: 'src/index.js',
-	output: [
-		{ file: pkg.module, 'format': 'es' },
-		{ file: pkg.main, 'format': 'umd', name }
-	],
-	plugins: [
-		svelte({
-			emitCss: false //true
-		}),
-		commonjs(),
-		resolve(),
-		// import css from 'rollup-plugin-css-only';
-		// css({
-		// 	output: 'bundle.css'
-		// })
-	]
-};
