+++ src/lib/index.js
@@ -1,3 +1,4 @@
+// Reexport your entry components here
 import VirtualTable from './VirtualTable.svelte'
 
 export default VirtualTable
